+++ src/custom-code-helpers/string-array/templates/string-array-calls-wrapper/Rc4Template.ts
@@ -3,17 +3,11 @@
  */
 export function Rc4Template (): string {
     return `
-        const rc4 = function (str, key) {
-            let s = [], j = 0, x, res = '', newStr = '';
+        const {rc4FunctionName} = function (str, key) {
+            let s = [], j = 0, x, output = '';
            
             str = {atobFunctionName}(str);
                 
-            for (let k = 0, length = str.length; k < length; k++) {
-                newStr += '%' + ('00' + str.charCodeAt(k).toString(16)).slice(-2);
-            }
-        
-            str = decodeURIComponent(newStr);
-                    	     
             let i;
                     	        
             for (i = 0; i < 256; i++) {
@@ -36,10 +30,10 @@ export function Rc4Template (): string {
                 x = s[i];
                 s[i] = s[j];
                 s[j] = x;
-                res += String.fromCharCode(str.charCodeAt(y) ^ s[(s[i] + s[j]) % 256]);
+                output += String.fromCharCode(str.charCodeAt(y) ^ s[(s[i] + s[j]) % 256]);
             }
                       
-            return res;
+            return output;
         }
     `;
 }